requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ply==3.11
psycopg2==2.9.11
python-dotenv==1.1.1
Werkzeug==3.1.3
